Moved serde code for Direction to Rect
