+++ src/prodbg/viewdock/src/rect/serialize.rs
@@ -1,7 +1,7 @@
 extern crate serde;
-use super::Rect;
+use super::{Rect, Direction};
 
-// Serialization
+// Serialization of Rect
 
 impl serde::ser::Serialize for Rect {
     fn serialize<S>(&self, serializer: &mut S) -> Result<(), S::Error> where S: serde::ser::Serializer {
@@ -23,7 +23,7 @@ impl<'a> serde::ser::MapVisitor for RectMapVisitor<'a> {
     }
 }
 
-// Deserialization
+// Deserialization of Rect
 
 impl serde::Deserialize for Rect {
     fn deserialize<D>(deserializer: &mut D) -> Result<Rect, D::Error> where D: serde::de::Deserializer {
@@ -108,4 +108,78 @@ impl serde::Deserialize for RectField  {
 
             deserializer.deserialize(RectFieldVisitor)
         }
+}
+
+// Serialization of Direction
+
+impl serde::ser::Serialize for Direction {
+    fn serialize<S>(&self, serializer: &mut S) -> Result<(), S::Error> where S: serde::ser::Serializer {
+        match *self {
+            Direction::Vertical => serde::ser::Serializer::serialize_unit_variant(serializer, "Direction", 0usize, "Vertical"),
+            Direction::Horizontal => serde::ser::Serializer::serialize_unit_variant(serializer, "Direction", 1usize, "Horizontal"),
+        }
+    }
+}
+
+// Deserialization of Direction
+
+impl serde::Deserialize for Direction {
+    fn deserialize<D>(deserializer: &mut D) -> Result<Direction, D::Error> where D: serde::de::Deserializer {
+        const VARIANTS: &'static [&'static str] = &["Vertical", "Horizontal"];
+        deserializer.deserialize_enum("Direction", VARIANTS, DirectionVisitor)
+    }
+}
+
+struct DirectionVisitor;
+
+impl serde::de::EnumVisitor for DirectionVisitor {
+    type Value = Direction;
+
+    fn visit<V>(&mut self, mut visitor: V) -> Result<Direction, V::Error> where V: serde::de::VariantVisitor {
+        match try!(visitor.visit_variant()) {
+            DirectionField::Vertical => {
+                try!(visitor.visit_unit());
+                Ok(Direction::Vertical)
+            },
+            DirectionField::Horizontal => {
+                try!(visitor.visit_unit());
+                Ok(Direction::Horizontal)
+            }
+        }
+    }
+}
+
+enum DirectionField {
+    Vertical,
+    Horizontal,
+}
+
+impl serde::Deserialize for DirectionField  {
+    fn deserialize<D>(deserializer: &mut D) -> Result<DirectionField, D::Error> where D: serde::de::Deserializer {
+        struct DirectionFieldVisitor;
+
+        impl serde::de::Visitor for DirectionFieldVisitor {
+            type Value = DirectionField;
+
+            fn visit_usize<E>(&mut self, value: usize) -> Result<DirectionField, E>
+                where E: serde::de::Error {
+                    match value {
+                        0usize => Ok(DirectionField::Vertical),
+                        1usize => Ok(DirectionField::Horizontal),
+                        _ => Err(serde::de::Error::invalid_value("expected a variant")),
+                    }
+                }
+
+            fn visit_str<E>(&mut self, value: &str) -> Result<DirectionField, E>
+                where E: serde::de::Error {
+                    match value {
+                        "Vertical" => Ok(DirectionField::Vertical),
+                        "Horizontal" => Ok(DirectionField::Horizontal),
+                        _ => Err(serde::de::Error::invalid_value("expected a variant")),
+                    }
+                }
+        }
+
+        deserializer.deserialize_struct_field(DirectionFieldVisitor)
+    }
 }
